requirements.txt
Django>=4.2.3
django-crispy-forms>=2.0
crispy-bootstrap5>=2024.2
django-filter>=23.2
django-import-export>=3.2.0
django-tables2>=2.6.0
djangorestframework>=3.14.0
python-dotenv>=1.0.0